Add Long Qian's code for tags to post.html
This is "Step 4" in Long Qian's recipe for adding tags to a github-hosted jekyll page.

NOTES:
To display the tags of a post inside the post page, we must modify the format of posts (file in _layouts/post).

For example, Long Qian suggests inserting the following code into the post layout (add your own style to it):

<span>[
  {% for tag in page.tags %}
    {% capture tag_name %}{{ tag }}{% endcapture %}
    <a href="/tag/{{ tag_name }}"><code class="highligher-rouge"><nobr>{{ tag_name }}</nobr></code>&nbsp;</a>
  {% endfor %}
]</span>

Note from Long Qian: a link is inserted to each of the tags of the post, at /tag/tag_name. We will come to that in the next step.