Long Qian's page layout for tagpages
This is "Step 5.1" from Long Qian's recipe for adding tags to a github-hosted jekyll page.

Long Qian's Notes:

The tag page needs to be generated for all tags, because no plugin can be used to generate it when Github is building the page. Therefore, we treat the tag page as a kind of normal Jekyll page, with the only difference that it uses another layout.

* THIS FILE IS the layout of tag pages. 
* IT MUST BE INCLUDED IN _layouts/ 
* CALL IT:  tagpage.html:

Long Qian's code will generate all tag pages, which will display all the posts that has the given tag, including post title, date and description.